requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.30
asyncpg==0.29.0
apscheduler==3.10.4
redis==5.0.4
cachetools==5.3.3
python-dotenv==1.0.1

